Safari-609.1.11.1/LayoutTests/http/tests/security/xssAuditor/anchor-url-dom-write-location-expected.txt

CONSOLE MESSAGE: line 6: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-dom-write-location.html?#<script>alert(String.fromCharCode(0x58,0x53,0x53))</script>' because its source code was found within the request. The auditor was enabled because the server did not send an 'X-XSS-Protection' header.
